requirements.txt
streamlit==1.29.0
requests==2.31.0
python-dotenv==1.0.0
PyGithub==1.59.1